requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiohttp-retry==2.8.3
aiosignal==1.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
asyncssh==2.12.0
asynctest==0.13.0
atpublic==3.0.1
attrs==22.1.0
boto3==1.24.51
botocore==1.27.51
cached-property==1.5.2
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.4
cycler==0.11.0
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dpath==2.0.6
dulwich==0.20.45
dvc==2.10.2
dvc-render==0.0.5
dvclive==0.9.0
flatbuffers==1.12
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2022.7.1
ftfy==6.1.1
funcy==1.17
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grandalf==0.6
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.structures==2.1.0
jaraco.text==3.8.1
jaraco.ui==2.3.0
jaraco.windows==5.7.0
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
mailchecker==4.1.19
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
more-itertools==8.14.0
multidict==6.0.2
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
path==16.4.0
pathspec==0.9.0
phonenumbers==8.12.53
Pillow==9.2.0
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pygit2==1.10.0
Pygments==2.12.0
pygtrie==2.5.0
pyparsing==3.0.9
python-benedict==0.25.2
python-dateutil==2.8.2
python-fsutil==0.6.1
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scmrepo==0.0.19
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
smmap==5.0.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
voluptuous==0.13.1
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.1

#local setup.py installation
-e .